refactor: move init-authoring helpers into the backend
The scaffold + repo-detection + headless-Claude-run helpers that `webmux init`
uses (buildStarterTemplate, detectInitProjectContext, buildInitPromptSpec,
buildInitAgentCommand, runInitAgentCommand, stream parsing) move from
bin/src/init-helpers.ts to backend/src/services/init-authoring.ts so the
server can drive the same flow for an upcoming "set up project on add"
feature. `bin/src/init.ts` now imports them from the backend.

Generic process/git/repo primitives (run, which, getGitRoot, detectProjectName,
RunResult) move to backend/src/lib/shell.ts and are re-exported from
bin/src/shared.ts, so every existing `./shared` consumer is unchanged and the
backend can share them without a backend→bin import. Test moves alongside the
code into backend/src/__tests__.

No behavior change.
